Fix: Price의 데이터를 usePriceState 외부로 분리
